Fix apply with multiple lines in .gitlayer
This invocation of xargs results in a command-line that looks like the
following:

git-layer repo1 dir1 repo2 dir2

The desired behaviour is to split this into three separate invocations,
and the easy solution is to use the -n flag to xargs, which sets the
maximum number of arguments before it splits the command into multiple
invocations.

So passing "-n 2" turns this into

git-layer repo1 dir1
git-layer repo2 dir2
...

Which is the desired behaviour.